Move step and sidecar replacements to new package
This commit moves logic related to variable replacement in Steps and Sidecars
out of pkg/apis into its own package. This will prevent this code from being
duplicated in the v1 api package. No functional changes.
