+++ pkg/container/container_replacements.go
@@ -14,22 +14,23 @@
  limitations under the License.
 */
 
-package v1beta1
+package container
 
 import (
+	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	"github.com/tektoncd/pipeline/pkg/substitution"
 	corev1 "k8s.io/api/core/v1"
 )
 
 // applyStepReplacements returns a StepContainer with variable interpolation applied.
-func applyStepReplacements(step *Step, stringReplacements map[string]string, arrayReplacements map[string][]string) {
+func applyStepReplacements(step *v1beta1.Step, stringReplacements map[string]string, arrayReplacements map[string][]string) {
 	c := step.ToK8sContainer()
 	applyContainerReplacements(c, stringReplacements, arrayReplacements)
 	step.SetContainerFields(*c)
 }
 
 // applySidecarReplacements returns a SidecarContainer with variable interpolation applied.
-func applySidecarReplacements(sidecar *Sidecar, stringReplacements map[string]string, arrayReplacements map[string][]string) {
+func applySidecarReplacements(sidecar *v1beta1.Sidecar, stringReplacements map[string]string, arrayReplacements map[string][]string) {
 	c := sidecar.ToK8sContainer()
 	applyContainerReplacements(c, stringReplacements, arrayReplacements)
 	sidecar.SetContainerFields(*c)
